extract unicorn.go to some files
